Change summary:     1. Updated ultrasonic sensor distance MIA values to from 10,12,11 to 0,0,0     2. Added MIA internal LED indicator logic for driver and motor     3. Moved ultrasonic code from 1Hz to 10Hz and updated MIA counter from 1000 to 100     4. Updated DBC file to make sure sensor_sonar distance values are in float rather than int16_t     5. Changed ultrasonic pin num values     6. Removed the use of ultrasonic_distance_s struct     7. Updated test cases

+++ projects/canster/l5_application/hardware/ultrasonic/ultrasonic.c
@@ -60,7 +60,7 @@ void ultrasonic__calculate_distance_from_obstacle(ultrasonic_sensor_s *u_sensor)
                distance_cm_conversion_factor);
 
 #if ULTRASONIC_DEBUG == 1
-  printf("Distance from obstacle = %lf cm\r\n", u_sensor->distance_from_obstacle);
+  printf("Distance from obstacle = %lf cm\r\n", (double)u_sensor->distance_from_obstacle);
 #endif
 }
